PM agent/requirements.txt

# Backend Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0

# Google Cloud & AI (The new compute engine)
langchain==0.1.9
langchain-google-vertexai==0.0.5
google-cloud-aiplatform==1.42.0

# Orchestration
langgraph==0.0.24
langchain-core==0.1.26

# Data & Persistence
chromadb==0.4.22
# Note: Chroma still needs a lightweight local build, but we won't use heavy local embeddings
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
aiofiles==23.2.1